fix: repair the historical legacy xiaomi path and information references (#572)
