+++ README.md
@@ -32,7 +32,7 @@
 
 
 ## What is Apache OzHera(Incubating)?
-Apache OzHera(Incubating) is an Application Observable Platform in the Cloud Native Era. open-sourced by the R&D efficiency team of Xiaomi-China region. Centered around applications, it integrates metrics monitoring, link tracing, logging, and alerts. It achieves a seamless connection and interaction from metrics to tracing to logging. Moreover, Apache OzHera(Incubating) offers a comprehensive monitoring dashboard featuring application health status lists, application metric boards, interface overviews, application overviews, gateway overviews, etc., along with clear and concise visual text alerts, enabling users to pinpoint issues accurately and efficiently.
+Apache OzHera(Incubating) is an Application Observable Platform in the Cloud Native Era. Centered around applications, it integrates metrics monitoring, link tracing, logging, and alerts. It achieves a seamless connection and interaction from metrics to tracing to logging. Moreover, Apache OzHera(Incubating) offers a comprehensive monitoring dashboard featuring application health status lists, application metric boards, interface overviews, application overviews, gateway overviews, etc., along with clear and concise visual text alerts, enabling users to pinpoint issues accurately and efficiently.
 
 ---
