+++ ozhera-demo-server/ozhera-demo-server-server/src/main/proto/MyService.proto
@@ -19,10 +19,10 @@
 
 syntax = "proto3";
 
-package com.xiaomi.mone.hera.demo.grpc;
+package org.apache.ozhera.demo.grpc;
 
 option java_multiple_files = true;
-option java_package = "com.xiaomi.mone.hera.demo.grpc";
+option java_package = "org.apache.ozhera.demo.grpc";
 option java_outer_classname = "HelloWorldProto";
 
 // The greeting service definition.
