+++ trace-etl/trace-etl-server/src/test/java/org/apache/ozhera/trace/etl/test/FastWriter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.xiaomi.youpin.trace.etl.test;
+package org.apache.ozhera.trace.etl.test;
 
 import org.apache.commons.lang3.tuple.Pair;
