+++ trace-etl/trace-etl-server/src/test/java/org/apache/ozhera/trace/etl/test/TestConsumer.java
@@ -17,8 +17,12 @@
  * under the License.
  */
 
-package com.xiaomi.youpin.trace.etl.test;
+package org.apache.ozhera.trace.etl.test;
 
-
-public class Test {
+/**
+ * @Description
+ * @Author dingtao
+ * @Date 2022/10/24 3:35 下午
+ */
+public class TestConsumer {
 }
