+++ trace-etl/trace-etl-service/src/test/java/org/apache/ozhera/trace/etl/test/Test.java
@@ -17,24 +17,8 @@
  * under the License.
  */
 
-package com.xiaomi.hera.trace.etl.es;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-import java.util.concurrent.atomic.AtomicInteger;
-
-/**
- * @author goodjava@qq.com
- * @date 2023/9/7 14:05
- */
-public class AtomicTest {
+package org.apache.ozhera.trace.etl.test;
 
 
-    @Test
-    public void testGetAndUpdate() {
-        AtomicInteger ai = new AtomicInteger(1);
-        Assert.assertEquals(1,ai.getAndUpdate(v->12));
-    }
-
+public class Test {
 }
